# contracts/openapi/tab-v1.yaml
openapi: 3.0.3
info:
  title: UTab API v1
  version: 0.1.1
  description: >
    Contract-first API used by tournament operations. This is the initial
    machine-readable contract for Phase 0.
servers:
  - url: https://tab.example.com
security:
  - sessionAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/privacy/erasure-requests:
    get:
      tags: [Privacy]
      summary: List erasure requests for one tournament
      operationId: listErasureRequests
      parameters:
        - $ref: "#/components/parameters/TournamentIdQuery"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [requested, approved, rejected, running, completed, failed, cancelled]
        - name: targetType
          in: query
          required: false
          schema:
            type: string
            enum: [speaker, adjudicator]
        - name: targetId
          in: query
          required: false
          schema:
            type: string
        - name: requestedBy
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 500
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeErasureRequestList"
    post:
      tags: [Privacy]
      summary: Create erasure request
      operationId: createErasureRequest
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureRequestCreateBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeErasureRequest"
  /api/v1/privacy/erasure-requests/{id}/approve:
    patch:
      tags: [Privacy]
      summary: Approve erasure request
      operationId: approveErasureRequest
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureRequestApproveBody"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeErasureRequest"
        "404":
          description: Erasure request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
        "409":
          description: Erasure request is not in a valid state for approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
  /api/v1/privacy/erasure-requests/{id}/reject:
    patch:
      tags: [Privacy]
      summary: Reject erasure request
      operationId: rejectErasureRequest
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureRequestDecisionBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeErasureRequest"
        "404":
          description: Erasure request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
        "409":
          description: Erasure request is not in a valid state for rejection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
  /api/v1/privacy/erasure-requests/{id}/cancel:
    patch:
      tags: [Privacy]
      summary: Cancel erasure request
      operationId: cancelErasureRequest
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureRequestDecisionBody"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeErasureRequest"
        "404":
          description: Erasure request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
        "409":
          description: Erasure request is not cancellable in current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
  /api/v1/privacy/erasure-requests/{id}/execute:
    post:
      tags: [Privacy]
      summary: Execute approved erasure request
      operationId: executeErasureRequest
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureRequestExecuteBody"
      responses:
        "200":
          description: Executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeErasureRequest"
        "404":
          description: Erasure request or target not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
        "409":
          description: Erasure request must be approved before execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKeyHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: Required for service-account authenticated POST/PATCH/DELETE requests.
...
    TournamentIdQuery:
      name: tournamentId
      in: query
      required: true
      schema:
        type: string
      description: Tournament ObjectId
  schemas:
...
    EnvelopeErasureRequest:
      type: object
      required: [data, errors]
      properties:
        data:
          $ref: "#/components/schemas/ErasureRequestDocument"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
    EnvelopeErasureRequestList:
      type: object
      required: [data, errors]
      properties:
        data:
          type: object
          required: [items, limit, hasMore, nextCursor]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ErasureRequestDocument"
            limit:
              type: integer
            hasMore:
              type: boolean
            nextCursor:
              type: string
              nullable: true
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
    EnvelopeError:
      type: object
      required: [data, errors]
      properties:
        data:
          nullable: true
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
...
    ErasureResult:
      type: object
      required: [tournamentId, entityType, entityId, redacted]
      properties:
        tournamentId:
          type: string
        entityType:
          type: string
          enum: [speaker, adjudicator]
        entityId:
          type: string
        redacted:
          type: boolean
        eraseMode:
          type: string
          enum: [anonymize, hard_delete]
        reason:
          type: string
        approvedBy:
          type: string
          nullable: true
        targetRefs:
          type: array
          items:
            type: string
        submissionCommentsCleared:
          type: integer
          minimum: 0
...
    ErasureRequestCreateBody:
      type: object
      required: [tournamentId, targetType, targetId, reason]
      properties:
        tournamentId:
          type: string
        targetType:
          type: string
          enum: [speaker, adjudicator]
        targetId:
          type: string
        reason:
          type: string
          minLength: 5
          maxLength: 500
        targetRefs:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 256
        eraseMode:
          type: string
          enum: [anonymize, hard_delete]
    ErasureRequestApproveBody:
      type: object
      required: [tournamentId]
      properties:
        tournamentId:
          type: string
        approvedBy:
          type: string
          minLength: 1
          maxLength: 128
        reauthPassword:
          type: string
          minLength: 1
          maxLength: 200
          description: >
            Required for interactive organizer flows. Service-account callers
            are exempt.
    ErasureRequestDecisionBody:
      type: object
      required: [tournamentId]
      properties:
        tournamentId:
          type: string
        reason:
          type: string
          minLength: 1
          maxLength: 500
        reauthPassword:
          type: string
          minLength: 1
          maxLength: 200
          description: >
            Required for interactive organizer flows. Service-account callers
            are exempt.
    ErasureRequestExecuteBody:
      type: object
      required: [tournamentId]
      properties:
        tournamentId:
          type: string
        reauthPassword:
          type: string
          minLength: 1
          maxLength: 200
          description: >
            Required for interactive organizer flows. Service-account callers
            are exempt.
    ErasureRequestDocument:
      type: object
      properties:
        _id:
          type: string
        tournamentId:
          type: string
        targetType:
          type: string
          enum: [speaker, adjudicator]
        targetId:
          type: string
        eraseMode:
          type: string
          enum: [anonymize, hard_delete]
        reason:
          type: string
        targetRefs:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [requested, approved, rejected, running, completed, failed, cancelled]
        requestedBy:
          type: string
        requestedAt:
          type: string
          format: date-time
        approvedBy:
          type: string
        approvedAt:
          type: string
          format: date-time
        rejectedBy:
          type: string
        rejectedAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
        executedBy:
          type: string
        executedAt:
          type: string
          format: date-time
        errorMessage:
          type: string
        result:
          $ref: "#/components/schemas/ErasureResult"
    ErrorItem:
      type: object
      required: [name, message]
      properties:
        name:
          type: string
        message:
          type: string